+++ app/src/main/java/com/battlelancer/seriesguide/movies/MoviesNowFragment.kt
@@ -305,6 +305,6 @@ class MoviesNowFragment : Fragment() {
         }
 
     companion object {
-        const val liftOnScrollTargetViewId = R.id.recyclerViewNow
+        val liftOnScrollTargetViewId = R.id.recyclerViewNow
     }
 }