+++ app/src/main/java/com/battlelancer/seriesguide/shows/ShowsAdapter.kt
@@ -173,7 +173,7 @@ class ShowsAdapter(
 
                 val timeAndNetwork = TextTools.dotSeparate(
                     network,
-                    releaseTimeShow?.formatWithDeviceZoneToDayAndTime(context, weekDay)
+                    releaseTimeShow?.formatWithDeviceZoneToDayAndTime()
                 )
 
                 return ShowItem(